Implement snake_body and prev_segment
A segment is added to the snake when the snake consumes food. snake_body
contains the coordinates for each snake segment, starting at the snake's
head. This list allows for easy manipulation of the snake's coordinates,
making rendering easier

prev_segment stores previous head location and is used as the value for
the next segment whenever the snake consumes food